refactor: change contextcontext to hoc withContext
